--- cmd/shortenertest/iteration5_test.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"testing"
-)
-
-// TestIteration5 is a blank test function
-func TestIteration5(t *testing.T) {
-	t.Skipf("Run iterations 1 and 4 manually with desired env vars configuration")
-}
