--- cmd/shortenertest/iteration7_test.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"testing"
-)
-
-// TestIteration7 is a blank test function
-func TestIteration7(t *testing.T) {
-	t.Skipf("Run iterations 1 and 4 manually with desired flags configuration")
-}
